requirements.txt
groq
gradio
requests
python-dotenv
transformers
torch 
torchaudio
PyPDF2
python-docx
python-pptx
pandas
openpyxl
streamlit